Fixed port to be 16 bits because I'm a ding dong

+++ cpp/include/Sockets/Endpoint.h
@@ -7,7 +7,7 @@ class Endpoint
 {
 public:
     std::string ipAddress;
-    uint8_t portNumber;
+    uint16_t portNumber;
 };
 
 #endif
